+++ inst/NEWS.md
@@ -1,4 +1,4 @@
-# CHANGES IN VERSION 2.14.05
+# CHANGES IN VERSION 2.17.0
 (GitHub master branch)
 
 ## BUG FIXES
@@ -7,6 +7,7 @@
 - Bug fix in `oncoplot` for duplicated values in gene list. [889](https://github.com/PoisonAlien/maftools/issues/889)
 
 ## ENHANCEMENTS
+- Added `showOnlyPathway` argument to `oncoplot`
 - Added `pathdb` argument to `PlotOncogenicPathways`. Issue: [923](https://github.com/PoisonAlien/maftools/issues/923)
 - Emit warnings when fishers test can not be performed during `somaticInteractions`. Issue: [921](https://github.com/PoisonAlien/maftools/issues/921)
 - Added `leftMar` and `topMar` arguments to `somaticInteractions`. Issue: [913](https://github.com/PoisonAlien/maftools/issues/913)
@@ -15,6 +16,9 @@
 - Added `sortByM1` and `sortByM2` argument to `coOncoplot`. Issue: [888](https://github.com/PoisonAlien/maftools/issues/888)
 - Added arguments `leftBarVline`, `leftBarVlineCol`, `rightBarVline`, `rightBarVlineCol` `topBarHline` `topBarHlineCol` to `oncoplot`. Issue: [874](https://github.com/PoisonAlien/maftools/issues/874)
 - Added `revPal` argument to `somaticInteractions`. Issue: [859](https://github.com/PoisonAlien/maftools/issues/859)
+
+# CHANGES IN VERSION 2.16.0
+(Bioconductor release version)
 
 # CHANGES IN VERSION 2.14.0
 (Bioconductor release version)